chore(ilp): refactored code to remove some spaghetti
